[chore #160406862] rename validations function to utils
